Make project compilable. Provide dummy implementations. Implement PriceCalculator

+++ src/Domain/Money.h
@@ -1,11 +1,13 @@
 #pragma once
 
+#include <cstdint>
+
 namespace Vertaler::ParkingSystem::Domain
 {
 
 struct Money
 {
-  int32_t amount = 0;
+  size_t amount = 0;
 };
 
 }// namespace Vertaler::ParkingSystem::Domain
